+++ src/Gibbs/GibbsMMulti.cpp
@@ -279,7 +279,7 @@ void GibbsMMulti::_tableStore(int mode, const cs* A)
 
   double distmax = 1.5 * model->getCova(0)->getRange();
   double dx = db->getDX(0);
-  int npas = ceil(distmax / dx);
+  int npas = (int) ceil(distmax / dx);
   Table tabmod(npas,3);
 
   // Retrieve the elements from the sparse matrix
@@ -309,7 +309,7 @@ void GibbsMMulti::_tableStore(int mode, const cs* A)
       ecr++;
 
       // Store the contribution to the covariance model
-      int ipas = floor(dist / dx + 0.5);
+      int ipas = (int) floor(dist / dx + 0.5);
       if (ipas >= 0 && ipas < npas)
       {
         tabmod.increment(ipas, 0, 1.);
